feat(api): routes endpoints fixes #33

+++ modules/services/api/src/edu/controllers/roles/roles.controller.ts
@@ -3,13 +3,14 @@ import { InjectMapper } from '@automapper/nestjs';
 import {
   Body,
   Controller,
+  ForbiddenException,
   NotFoundException,
   Param,
   ParseUUIDPipe,
   Query,
   UseGuards,
 } from '@nestjs/common';
-import { ApiTags } from '@nestjs/swagger';
+import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
 import { UserWithPermissions } from '@vidya/api/auth/decorators';
 import { AuthenticatedUser } from '@vidya/api/auth/guards';
 import { UserPermissions } from '@vidya/api/auth/utils';
@@ -17,13 +18,14 @@ import * as dto from '@vidya/api/edu/dto';
 import { RoleExistsPipe } from '@vidya/api/edu/pipes';
 import { RolesService } from '@vidya/api/edu/services';
 import { CrudDecorators } from '@vidya/api/utils';
+import * as domain from '@vidya/domain';
 import * as entities from '@vidya/entities';
 import { Routes } from '@vidya/protocol';
 
 const Crud = CrudDecorators({
   entityName: 'Role',
   getOneResponseDto: dto.GetRoleResponse,
-  getManyResponseDto: dto.GetRoleSummariesListResponse,
+  getManyResponseDto: dto.GetRolesResponse,
   createOneResponseDto: dto.CreateRoleResponse,
   updateOneResponseDto: dto.UpdateRoleResponse,
   deleteOneResponseDto: dto.DeleteRoleResponse,
@@ -31,6 +33,7 @@ const Crud = CrudDecorators({
 
 @Controller()
 @ApiTags('Roles')
+@ApiBearerAuth()
 @UseGuards(AuthenticatedUser)
 export class RolesController {
   constructor(
@@ -50,7 +53,7 @@ export class RolesController {
     if (roles.length === 0) {
       throw new NotFoundException(`Role with id ${id} not found`);
     }
-    return this.mapper.map(roles[0], entities.Role, dto.Role);
+    return this.mapper.map(roles[0], entities.Role, dto.RoleDetails);
   }
 
   /* -------------------------------------------------------------------------- */
@@ -61,13 +64,13 @@ export class RolesController {
   async getMany(
     @Query() query: dto.GetRoleSummariesListQuery,
     @UserWithPermissions() permissions: UserPermissions,
-  ): Promise<dto.GetRoleSummariesListResponse> {
+  ): Promise<dto.GetRolesResponse> {
     const roles = await this.rolesService.findPermittedBy(permissions, {
       organizationId: query.organizationId,
       schoolId: query.schoolId,
     });
-    return new dto.GetRoleSummariesListResponse({
-      roles: this.mapper.mapArray(roles, entities.Role, dto.RoleSummary),
+    return new dto.GetRolesResponse({
+      items: this.mapper.mapArray(roles, entities.Role, dto.RoleSummary),
     });
   }
 
@@ -78,11 +81,20 @@ export class RolesController {
   @Crud.CreateOne(Routes().edu.roles.create())
   async createOne(
     @Body() request: dto.CreateRoleRequest,
+    @UserWithPermissions() userPermissions: UserPermissions,
   ): Promise<dto.CreateRoleResponse> {
+    this.checkUserPermissions(
+      userPermissions,
+      ['roles:create'],
+      request.organizationId,
+      request.schoolId,
+    );
+
     const entity = await this.rolesService.create(
       this.mapper.map(request, dto.CreateRoleRequest, entities.Role),
     );
-    return new dto.CreateRoleResponse(entity.id);
+
+    return this.mapper.map(entity, entities.Role, dto.CreateRoleResponse);
   }
 
   /* -------------------------------------------------------------------------- */
@@ -93,12 +105,23 @@ export class RolesController {
   async updateOne(
     @Param('id', new ParseUUIDPipe(), RoleExistsPipe) id: string,
     @Body() request: dto.UpdateRoleRequest,
+    @UserWithPermissions() userPermissions: UserPermissions,
   ): Promise<dto.UpdateRoleResponse> {
-    await this.rolesService.updateOneBy(
+    // Check if user has permission to update role
+    let role = await this.rolesService.findOneBy({ id });
+    this.checkUserPermissions(
+      userPermissions,
+      ['roles:update'],
+      role.organizationId,
+      role.schoolId,
+    );
+
+    // User has permission to update role. Proceed with update.
+    role = await this.rolesService.updateOneBy(
       { id },
       this.mapper.map(request, dto.UpdateRoleRequest, entities.Role),
     );
-    return new dto.UpdateRoleResponse(id);
+    return this.mapper.map(role, entities.Role, dto.UpdateRoleResponse);
   }
 
   /* -------------------------------------------------------------------------- */
@@ -108,8 +131,52 @@ export class RolesController {
   @Crud.DeleteOne(Routes().edu.roles.delete(':id'))
   async deleteOne(
     @Param('id', new ParseUUIDPipe(), RoleExistsPipe) id: string,
+    @UserWithPermissions() userPermissions: UserPermissions,
   ): Promise<dto.DeleteRoleResponse> {
+    // Check if user has permission to delete role
+    const role = await this.rolesService.findOneBy({ id });
+    this.checkUserPermissions(
+      userPermissions,
+      ['roles:delete'],
+      role.organizationId,
+      role.schoolId,
+    );
+
+    // User has permission to delete role. Proceed with deletion.
     await this.rolesService.deleteOneBy({ id });
-    return new dto.DeleteRoleResponse(id);
+    return new dto.DeleteRoleResponse({ success: true });
+  }
+
+  /* -------------------------------------------------------------------------- */
+  /*                               Private methods                              */
+  /* -------------------------------------------------------------------------- */
+
+  // TODO !HIGH Extract to base ProtectedCrud controller
+  private checkUserPermissions(
+    userPermissions: UserPermissions,
+    requiredPermissions: domain.PermissionKey[],
+    organizationId: string,
+    schoolId?: string,
+  ) {
+    // Check if user has permission to create roles
+    // on the organization level
+    const permittedOnOrgLevel = userPermissions
+      .getPermittedOrganizations(requiredPermissions)
+      .includes(organizationId);
+
+    // Check if user has permission to create roles
+    // on the school level
+    const permittedOnSchoolLevel = schoolId
+      ? userPermissions
+          .getPermittedSchools(requiredPermissions)
+          .includes(schoolId)
+      : true;
+
+    // If user does not have permission to create roles
+    // on either the organization or school level, throw an error
+    // indicating that the user does not have permission
+    if (!permittedOnOrgLevel || !permittedOnSchoolLevel) {
+      throw new ForbiddenException('User does not have permission');
+    }
   }
 }
